requirements.txt
pycryptodome
pycryptodomex
rsa
hashlib
